+++ jpro-file/example/src/main/java/one/jpro/platform/file/example/editor/TextEditorSample.java
@@ -17,6 +17,7 @@ import javafx.stage.Stage;
 import one.jpro.platform.file.ExtensionFilter;
 import one.jpro.platform.file.FileSource;
 import one.jpro.platform.file.dropper.FileDropper;
+import one.jpro.platform.file.picker.DirectoryOpenPicker;
 import one.jpro.platform.file.picker.FileOpenPicker;
 import one.jpro.platform.file.picker.FileSavePicker;
 import org.apache.commons.io.FilenameUtils;
@@ -88,6 +89,7 @@ public class TextEditorSample extends Application {
 
         FileDropper fileDropper = FileDropper.create(contentPane);
         fileDropper.setExtensionFilter(SUBTITLE_EXTENSION_FILTER);
+        // Nah?
         fileDropper.setOnDragEntered(event -> {
             dropPane.pseudoClassStateChanged(FILES_DRAG_OVER_PSEUDO_CLASS, true);
             contentPane.getChildren().setAll(textArea, dropPane);
@@ -115,9 +117,18 @@ public class TextEditorSample extends Application {
             contentPane.getChildren().setAll(textArea);
         });
 
+        Button openDirectory = new Button("Open Directory", new FontIcon(Material2AL.FOLDER_OPEN));
+        DirectoryOpenPicker directoryOpenPicker = DirectoryOpenPicker.create(openDirectory);
+        directoryOpenPicker.setOnFilesSelected(fileSources -> {
+            fileSources.stream().findFirst().ifPresent(fileSource -> {
+                // Note: We only test whether choosing a directory works
+                System.out.println("Chosen directory: " + fileSource.uploadFileAsync().join().getAbsolutePath());
+            });
+        });
+
         Region spacer = new Region();
         HBox.setHgrow(spacer, Priority.ALWAYS);
-        HBox controlsBox = new HBox(newButton, openButton, spacer);
+        HBox controlsBox = new HBox(newButton, openButton, openDirectory, spacer);
         controlsBox.getStyleClass().add("controls-box");
 
         final Button saveAsButton;